Switch to readFileSync vs ReadFile
ReadFile keeps the files open and can lead to hitting file
limits. Switching to readFileSync avoids this. And since this is not
a web request this should be safe (not blocking).

Also broke up some code to help make it more digestable. There
appears to be potential to keep a cache of file dependencies for
memoization to
avoid traversing duplicates.